requirements.txt
django
Markdown
django-filter
djangorestframework
django-ckeditor
Pillow